refactor(checkstyle): extract checkstyle
